+++ scanner/src/cmd.rs
@@ -2,9 +2,9 @@ use std::path::PathBuf;
 
 use argh::{FromArgValue, FromArgs};
 
-pub struct Target(pub usize);
+pub struct Address(pub usize);
 
-impl FromArgValue for Target {
+impl FromArgValue for Address {
     fn from_arg_value(value: &str) -> Result<Self, String> {
         let value = value.trim_start_matches("0x");
         let address = usize::from_str_radix(value, 16).map_err(|e| e.to_string())?;
@@ -24,7 +24,7 @@ impl FromArgValue for Offset {
 }
 
 #[derive(FromArgs)]
-#[argh(description = "Commands.")]
+#[argh(description = "PointerSearch-X")]
 pub struct Commands {
     #[argh(subcommand)]
     pub cmds: CommandEnum,
@@ -33,17 +33,22 @@ pub struct Commands {
 #[derive(FromArgs)]
 #[argh(subcommand)]
 pub enum CommandEnum {
-    Scanner(SubCommandScan),
+    Scan1(SubCommandScan1),
+    Scan2(SubCommandScan2),
     Diff(SubCommandDiff),
 }
 
 #[derive(FromArgs)]
-#[argh(subcommand, name = "scan", description = "scanner")]
-pub struct SubCommandScan {
+#[argh(
+    subcommand,
+    name = "s1",
+    description = "Scan mode 1, select some modules to set as base addresses."
+)]
+pub struct SubCommandScan1 {
     #[argh(option, short = 'f', description = "ptrs file path")]
     pub file: PathBuf,
     #[argh(option, short = 't', description = "target address")]
-    pub target: Target,
+    pub target: Address,
     #[argh(option, default = "7", short = 'd', description = "depth")]
     pub depth: usize,
     #[argh(option, default = "Offset((0, 600))", short = 'o', description = "offset")]
@@ -55,7 +60,30 @@ pub struct SubCommandScan {
 }
 
 #[derive(FromArgs)]
-#[argh(subcommand, name = "diff", description = "diff")]
+#[argh(subcommand, name = "s2", description = "Scan mode 2, set address as the base address.")]
+pub struct SubCommandScan2 {
+    #[argh(option, short = 'f', description = "ptrs file path")]
+    pub file: PathBuf,
+    #[argh(option, short = 's', description = "start address")]
+    pub start: Address,
+    #[argh(option, short = 't', description = "target address")]
+    pub target: Address,
+    #[argh(option, default = "7", short = 'd', description = "depth")]
+    pub depth: usize,
+    #[argh(option, default = "Offset((0, 600))", short = 'o', description = "offset")]
+    pub offset: Offset,
+    #[argh(option, default = "3", short = 'n', description = "node")]
+    pub node: usize,
+    #[argh(option, description = "out dir")]
+    pub dir: Option<PathBuf>,
+}
+
+#[derive(FromArgs)]
+#[argh(
+    subcommand,
+    name = "diff",
+    description = "Compare and get the intersecting parts of two .scandata files."
+)]
 pub struct SubCommandDiff {
     #[argh(option, description = "file1 path")]
     pub f1: PathBuf,