+++ scanner/src/scan.rs
@@ -2,11 +2,18 @@ use std::{fs::OpenOptions, io::BufWriter, path::Path};
 
 use ptrsx::{Params, PtrsxScanner};
 
-use super::{select_base_module, Error, Spinner, SubCommandScan};
+use super::{select_base_module, Address, Error, Offset, Spinner, SubCommandScan1, SubCommandScan2};
 
-impl SubCommandScan {
+impl SubCommandScan1 {
     pub fn init(self) -> Result<(), Error> {
-        let SubCommandScan { ref file, target, depth, offset, node, dir } = self;
+        let Self {
+            ref file,
+            target: Address(target),
+            depth,
+            offset: Offset(offset),
+            node,
+            dir,
+        } = self;
 
         if depth <= node {
             println!("Error: depth must be greater than node. current depth({depth}), node({node}).")
@@ -26,7 +33,7 @@ impl SubCommandScan {
         pages
             .iter()
             .map(|m| (m.start, m.path, ptrsx.range_address(m).collect::<Vec<_>>()))
-            .try_for_each(|(base, name, points)| {
+            .try_for_each(|(base, name, ref points)| {
                 let name = Path::new(name)
                     .file_name()
                     .and_then(|f| f.to_str())
@@ -37,17 +44,56 @@ impl SubCommandScan {
                     .append(true)
                     .create_new(true)
                     .open(file)?;
+                #[rustfmt::skip]
                 let params = Params {
-                    base,
-                    depth,
-                    target: target.0,
-                    node,
-                    range: offset.0,
-                    points: &points,
+                    base, depth, target, node, offset, points,
                     writer: &mut BufWriter::new(file),
                 };
                 ptrsx.scan(params)
             })?;
+        spinner.stop("Pointer chain is scanned.");
+
+        Ok(())
+    }
+}
+
+impl SubCommandScan2 {
+    pub fn init(self) -> Result<(), Error> {
+        let Self {
+            ref file,
+            start: Address(start),
+            target: Address(target),
+            depth,
+            offset: Offset(offset),
+            node,
+            dir,
+        } = self;
+        if depth <= node {
+            println!("Error: depth must be greater than node. current depth({depth}), node({node}).")
+        }
+
+        let mut spinner = Spinner::start("Start loading cache...");
+        let ptrsx = PtrsxScanner::load_with_file(file)?;
+        spinner.stop("cache loaded.");
+        let dir = dir.unwrap_or_default();
+
+        let mut spinner = Spinner::start("Start scanning pointer chain...");
+        let file = dir.join(format!("{:#x}.scandata", start));
+        let file = OpenOptions::new()
+            .write(true)
+            .append(true)
+            .create_new(true)
+            .open(file)?;
+        #[rustfmt::skip]
+        let params = Params {
+            base: 0,
+            depth, target, node, offset,
+            points: &[start],
+            writer: &mut BufWriter::new(file),
+        };
+
+        ptrsx.scan(params)?;
+
         spinner.stop("Pointer chain is scanned.");
 
         Ok(())